backend/requirements.txt
torch
torchvision
transformers
streamlit
sentence-transformers
faiss-cpu
pillow
opencv-python
numpy
